% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_targets.R
\name{check_targets}
\alias{check_targets}
\title{Title}
\usage{
check_targets(
  targets,
  seed = NULL,
  target.value.names = "value",
  seed.value.name = "value",
  max.error = 0.01
)
}
\arguments{
\item{max.error}{}
}
\description{
Title
}
