% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ia_cast.R
\name{ip_backcast}
\alias{ip_backcast}
\title{Title}
\usage{
ip_backcast(
  datatable,
  series_target = "value",
  series_base = "value_base",
  growth_over = c("Year")
)
}
\arguments{
\item{growth_over}{}
}
\description{
Title
}
