% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{ip_expand}
\alias{ip_expand}
\title{Title}
\usage{
ip_expand(
  targets,
  target.value.names = "value",
  names.exclude = c("value"),
  value.set = 1,
  value.name = "value",
  override.warning = FALSE,
  max.error = 0.01,
  max.iterations = 25,
  freeze_cells = NULL,
  freeze_cells.value.name = "value",
  freeze_slice = NULL,
  freeze_slice.value.names = "value",
  minmax_cells = NULL,
  minmax_cells.value.names = c("value_min", "value_max"),
  minmax_slice = NULL,
  minmax_slice.value.names = c("value_min", "value_max"),
  minmax.smash.param = 1/3,
  save.tars = FALSE,
  show.messages = TRUE
)
}
\arguments{
\item{show.messages}{}
}
\description{
Title
}
