% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_wrappers.R
\name{ip_fit_sl}
\alias{ip_fit_sl}
\title{Title}
\usage{
ip_fit_sl(
  datatable,
  targets,
  assumptions,
  datatable.value.name = "value",
  target.value.names = "value",
  assumption.value.names = c("value", "value_min", "value_max"),
  assumption.drop.names = c("Notes"),
  max.error = 0.01,
  max.iterations = 25,
  minmax.smash.param = 1/3,
  save.tars = FALSE,
  show.messages = TRUE
)
}
\arguments{
\item{show.messages}{}
}
\description{
Title
}
