% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{ip_growth_transform}
\alias{ip_growth_transform}
\title{Title}
\usage{
ip_growth_transform(
  datatable,
  target,
  series_start = "value",
  series_target = "growth",
  series_growth_over = "growth__over",
  sep = " + ",
  growth_distance = 1
)
}
\arguments{
\item{growth_distance}{}
}
\description{
Title
}
