% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_assumptions.R
\name{ip_load_assumptions}
\alias{ip_load_assumptions}
\title{Title}
\usage{
ip_load_assumptions(
  assumptions,
  freeze.name = "value",
  minmax.name = c("value_min", "value_max"),
  drop.names = c("Notes")
)
}
\arguments{
\item{drop.names}{}
}
\description{
Title
}
