% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_slice_a.R
\name{ip_load_slice_a}
\alias{ip_load_slice_a}
\title{Title}
\usage{
ip_load_slice_a(
  slice,
  slice.value.name = "value",
  drop.names = c("Notes"),
  prefix = "frz"
)
}
\arguments{
\item{prefix}{}
}
\description{
Title
}
