% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{ip_mask_seed}
\alias{ip_mask_seed}
\title{Apply a layer of multiplicative masks to a seed.}
\usage{
ip_mask_seed(seed, mask, mask.value.name = "mask", drop.names = c("Notes"))
}
\description{
Apply a layer of multiplicative masks to a seed.
}
