% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss.R
\name{ip_miss}
\alias{ip_miss}
\title{Title}
\usage{
ip_miss(
  datatable,
  target,
  series_start = "value",
  series_target = "value",
  series_type = "tar"
)
}
\arguments{
\item{series_type}{}
}
\description{
Title
}
