% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss.R
\name{ip_miss_a}
\alias{ip_miss_a}
\title{Compare data frame to target subtotals and calculate absolute error.}
\usage{
ip_miss_a(
  datatable,
  target_series,
  series_start = "value",
  series_target = "tar1",
  series_type = "tar"
)
}
\description{
Compare data frame to target subtotals and calculate absolute error.
}
