% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{ip_scale_a}
\alias{ip_scale_a}
\title{Single target scale function using single data frame.}
\usage{
ip_scale_a(
  datatable,
  target_series,
  series_start = "value",
  series_target = "tar1",
  series_type = "tar"
)
}
\description{
Single target scale function using single data frame.
}
