% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shares.R
\name{ip_shares_transform}
\alias{ip_shares_transform}
\title{Title}
\usage{
ip_shares_transform(
  datatable,
  target,
  series_start = "value",
  series_target = "share",
  series_share_of = "share__of",
  sep = " + "
)
}
\arguments{
\item{sep}{}
}
\description{
Title
}
